backend/requirements.txt
# Backend
fastapi==0.104.1
fastapi[all]==0.104.1
uvicorn[standard]==0.24.0
python-multipart==0.0.6
sqlalchemy==2.0.23
psycopg2-binary==2.9.9
alembic==1.12.1
python-jose[cryptography]==3.3.0
passlib[bcrypt]==1.7.4
python-dotenv==1.0.0
pydantic==2.5.0
pydantic-settings==2.1.0

# ML and NLP
torch==2.1.1
transformers==4.36.0
datasets==2.15.0
tokenizers==0.15.0
scikit-learn==1.3.2
numpy==1.24.3
pandas==2.1.4

# OCR and image processing
pytesseract==0.3.10
Pillow==10.1.0
opencv-python-headless==4.8.1.78
scikit-image==0.22.0

# Utilities
tqdm==4.66.1
matplotlib==3.8.2
seaborn==0.13.0 
